+++ internal/models/dto/user.go
@@ -1,5 +1,5 @@
 package dto
 
 type UserWallet struct {
-	UserID uint `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 }
