+++ internal/models/user.go
@@ -4,8 +4,8 @@ package models
 // @description: UserWallet is a wallet of a user.
 type UserWallet struct {
 	// @description: UserID is a unique identifier of the user, that owns this wallet.
-	UserID uint `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 
 	// @description: Balance is a current balance of the wallet.
-	Balance uint `json:"balance"`
+	Balance uint64 `json:"balance"`
 }
